feat(api): API overhaul
In this new version of DI, an instance of a DIContainer is no longer exported.
Instead, you must instantiate one yourself. This enables you to instantiate several, scoped service containers on multiple levels in your applications.
Additionally, there has been improvements to the API, as well as better error handling, and better documentation.